# Remove methods from core