--- a/symengine/basic-methods.inc
+++ b/symengine/basic-methods.inc
@@ -1,0 +1,2 @@
+#define SYMENGINE_INCLUDE_METHODS \
+virtual void accept(Visitor &v) const;