--- a/symengine/basic-methods.inc
+++ b/symengine/basic-methods.inc
@@ -1,2 +1,3 @@
 #define SYMENGINE_INCLUDE_METHODS \
-virtual void accept(Visitor &v) const;
+virtual void accept(Visitor &v) const; \
+virtual RCP<const Basic> diff(const RCP<const Symbol> &x) const;