--- a/symengine/basic-methods.inc
+++ b/symengine/basic-methods.inc
@@ -1,3 +1,3 @@
-#define SYMENGINE_INCLUDE_METHODS \
-virtual void accept(Visitor &v) const; \
-virtual RCP<const Basic> diff(const RCP<const Symbol> &x) const;
+#define SYMENGINE_INCLUDE_METHODS(SUFFIX) \
+virtual void accept(Visitor &v) const SUFFIX; \
+virtual RCP<const Basic> diff(const RCP<const Symbol> &x) const SUFFIX;