--- a/symengine/basic.h
+++ b/symengine/basic.h
@@ -256,11 +256,14 @@ void hash_combine(std::size_t& seed, const T& v);
 //! Inline members and functions
 #include "basic-inl.h"
 
+#include "basic-methods.inc"
+
 // Macro to define the type_code_id variable and its getter method
 #define IMPLEMENT_TYPEID(ID) \
 /*! Type_code_id shared by all instances */ \
 const static TypeID type_code_id = ID; \
 /*! Virtual function that gives the type_code_id of the object */ \
-virtual TypeID get_type_code() const { return type_code_id; };
+virtual TypeID get_type_code() const { return type_code_id; }; \
+SYMENGINE_INCLUDE_METHODS
 
 #endif