--- a/symengine/basic.h
+++ b/symengine/basic.h
@@ -145,8 +145,6 @@ public:
     std::string __str__() const;
 
     //! Returns the derivative of 'self' w.r.t Symbol `x`
-    virtual RCP<const Basic> diff(const RCP<const Symbol> &x) const;
-
     virtual RCP<const Basic> diff2(const RCP<const Symbol> &x) const = 0;
 
     //! Substitutes 'subs_dict' into 'self'.