--- a/symengine/complex.h
+++ b/symengine/complex.h
@@ -285,8 +285,6 @@ public:
     virtual RCP<const Number> rpow(const Number &other) const {
         throw std::runtime_error("Not implemented.");
     };
-
-    virtual void accept(Visitor &v) const;
 };
 
 } // SymEngine