--- a/symengine/complex_mpc.h
+++ b/symengine/complex_mpc.h
@@ -336,8 +336,6 @@ public:
             throw std::runtime_error("Not implemented.");
         }
     }
-
-    virtual void accept(Visitor &v) const;
 };
 
 inline RCP<const ComplexMPC> complex_mpc(mpc_class x) {