--- a/symengine/derivative.cpp
+++ b/symengine/derivative.cpp
@@ -15,6 +15,44 @@
 
 namespace SymEngine {
 
+class DiffImplementation {
+public:
+    static RCP<const Basic> diff(const Basic &self,
+            const RCP<const Symbol> &x) {
+        return Derivative::create(self.rcp_from_this(), {x});
+    }
+
+    static RCP<const Basic> diff(const Number &self,
+            const RCP<const Symbol> &x) {
+        return zero;
+    }
+
+    static RCP<const Basic> diff(const Sin &self,
+            const RCP<const Symbol> &x) {
+        return mul(cos(self.get_arg()), self.get_arg()->diff(x));
+    }
+
+    static RCP<const Basic> diff(const Cos &self,
+            const RCP<const Symbol> &x) {
+        return mul(mul(minus_one, sin(self.get_arg())),
+                self.get_arg()->diff(x));
+    }
+
+    // ...and so on for all the other functions that want to return a specific
+    // derivative, instead of the general diff(const Basic &, ...) one.
+};
+
+#define DIFF2(CLASS) \
+RCP<const Basic> CLASS::diff2(const RCP<const Symbol> &x) const { \
+    return DiffImplementation::diff(*this, x); \
+};
+
+
+#define SYMENGINE_ENUM(TypeID, Class) DIFF2(Class)
+#include "symengine/type_codes.inc"
+#undef SYMENGINE_ENUM
+
+
 #define IMPLEMENT_DIFF(CLASS) \
 RCP<const Basic> CLASS::diff(const RCP<const Symbol> &x) const \
 { \