--- a/symengine/derivative.cpp
+++ b/symengine/derivative.cpp
@@ -392,33 +392,4 @@ RCP<const Basic> CLASS::diff2(const RCP<const Symbol> &x) const { \
 #undef SYMENGINE_ENUM
 
 
-#define IMPLEMENT_DIFF(CLASS) \
-RCP<const Basic> CLASS::diff(const RCP<const Symbol> &x) const \
-{ \
-    return Derivative::create(rcp_from_this(), {x}); \
-}
-
-#define IMPLEMENT_DIFF0(CLASS) \
-RCP<const Basic> CLASS::diff(const RCP<const Symbol> &x) const \
-{ \
-    return zero; \
-}
-
-IMPLEMENT_DIFF(FunctionWrapper)
-IMPLEMENT_DIFF(UpperGamma)
-IMPLEMENT_DIFF(LowerGamma)
-IMPLEMENT_DIFF(Gamma)
-IMPLEMENT_DIFF(LeviCivita)
-IMPLEMENT_DIFF(KroneckerDelta)
-IMPLEMENT_DIFF(Dirichlet_eta)
-IMPLEMENT_DIFF(UnivariateSeries)
-
-IMPLEMENT_DIFF0(RealDouble)
-IMPLEMENT_DIFF0(ComplexDouble)
-IMPLEMENT_DIFF0(Complex)
-IMPLEMENT_DIFF0(Rational)
-IMPLEMENT_DIFF0(Integer)
-IMPLEMENT_DIFF0(NumberWrapper)
-
-
 } // SymEngine