… symengine/CMakeLists.txt
… symengine/functions.cpp
… symengine/mul.h
… symengine/number.cpp
… symengine/number.h
… symengine/polynomial.cpp
… symengine/polynomial.h
… symengine/pow.cpp
… symengine/pow.h
… symengine/rational.h
… symengine/real_double.h
… symengine/real_mpfr.h
… symengine/series.h
… symengine/symbol.cpp
… symengine/symbol.h