For now, I just removed `accept` and `diff`. This will work for everything, like `subs`, etc.

This works similar to the visitor pattern, except that only a single virtual function call is used.

By using the @isuruf's trick with the `bvisit` inline methods, this PR implements all derivatives as `DiffImplementation::diff` static methods.

Performance should not be affected by this PR.