@certik 2016-01-03
Now the only failure seems to be for optional MPFR and Arb. In fa1776303dd4e86fd071725664de794059781f06 I implemented visitor like approach, for a new method `diff2`. Once I implement all derivatives, I get rid of `diff` and rename `diff2` to `diff`. I posted my current status here, so that you can see the design. It's using static methods, but the `diff2` method can just as well accept an instance of `DiffImplementation` and call regular non-static methods on it, then the instance can save intermediate data inside it, just like the visitors work. Also since I use the type codes macros trick, this takes care of the optional MPFR, Arb, etc. issues.

So now the only difference between visitor pattern and this is that the visitor pattern allows the user to implement new functionality without touching the core at all, for the price of one extra virtual function call. This PR gets rid of the extra virtual function call, but it requires to add a new method to `basic-methods.inc` for each new algorithm (and thus full recompilation), but that's the only modification to our header files. The rest is then implemented in a .cpp file and thus is fully local.

And so I think we can use this approach to implement critical core functionality where we require maximum speed. We'll still allow the visitor approach, so that users can try out new things without modifying the core. We should also see if there is a way to unify the visitor class with the `DiffImplementation` class, so that one can (perhaps) just subclass a different class to switch back and forth between a visitor pattern and this new approach.

@isuruf and others, let me know what you think.