@certik 2016-01-04
I finished the implementation and removed all old diff methods and instead use the new ones. Tests pass on Travis, the Appveyor failures are unrelated (we should however fix them to make sure this PR builds with MSVC).

This is ready for review.