@isuruf 2016-01-04
@certik, there's another way to implement this with little change.

For each class we add methods like 
```
accept(StrPrinterFinal &v) {
     v.bvisit(*this);
}
```
where `StrPrinterFinal` is defined as follows.

```
class StrPrinterFinal : public BaseVisitor<StrPrinterFinal, StrPrinter> {
     using StrPrinter::bvisit;
}
```
and then it should work without other changes.

This way we keep the functionality of Visitors. Visitors with no specific `accept` methods will have a cost of two virtual function calls. Others will have a cost of only one virtual function call.