@certik 2016-01-04
@isuruf indeed, we can now do that too. Do you have any objections against merging this PR?

As to the Appveyor failure, I reported it here: http://help.appveyor.com/discussions/problems/3934-cannot-open-include-file-stdioh-no-such-file-or-directory