@isuruf 2016-01-04
Looks good to me. +1 to merge after tests pass