@certik 2016-01-04
I pushed in two more patches that implement your idea for `eval_double`. Indeed it works. I benchmarked with this patch:
```diff
diff --git a/benchmarks/eval_double1.cpp b/benchmarks/eval_double1.cpp
index 4bcda19..92e4451 100644
--- a/benchmarks/eval_double1.cpp
+++ b/benchmarks/eval_double1.cpp
@@ -42,7 +42,7 @@ int main(int argc, char* argv[])
 
     auto t1 = std::chrono::high_resolution_clock::now();
     for (int i = 0; i < 500; i++)
-        r = eval_double(*e);
+        r = eval_double_single_dispatch2(*e);
     auto t2 = std::chrono::high_resolution_clock::now();
     std::cout
         << std::chrono::duration_cast<std::chrono::milliseconds>(t2-t1).count()
```
without it I got:
```
certik@redhawk:~/repos/symengine/benchmarks(methods)$ ./eval_double1 
4620ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/symengine/benchmarks(methods)$ ./eval_double1 
4758ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/symengine/benchmarks(methods)$ ./eval_double1 
4773ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/symengine/benchmarks(methods)$ ./eval_double1 
4724ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
```
with it I got:
```
certik@redhawk:~/repos/symengine(methods)$ ./benchmarks/eval_double1 
4698ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/symengine(methods)$ ./benchmarks/eval_double1 
4680ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/symengine(methods)$ ./benchmarks/eval_double1 
4674ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/symengine(methods)$ ./benchmarks/eval_double1 
4594ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
```
So it seems it is a bit faster.

I first tried to overload `accept`, and it compiled, but it was still using the two virtual calls, because the `apply` method was at compile time dispatching to the general `accept`. This is fixed by implementing `accept` again in the `EvalRealDoubleVisitorFinal`. However, because it is so easy to get this wrong, I decided to rename the specialized `accept` to `accept_EvalRealDoubleVisitorFinal`, then there is no confusion about which method is being called.