@certik 2016-01-04
In 519e467b298ffe268ab70d4917992ce1c412c8c7 I ported `eval_complex_double()` to show how easy it is.

I am done with this PR. The rest we can do in new PRs. @isuruf are you ok with the new patches as well?