@certik 2016-01-04
Hm, but the `apply` calls implemented e.g. in `EvalDoubleVisitor::bvisit(const Pow &x)` still call `EvalDoubleVisitor::apply`, that calls `Basic::accept(Visitor &v)` which is slow...