@isuruf 2016-01-04
@certik, yes, let's first get 6963b8c in and then see how to improve