@certik 2016-01-04
Indeed. I reseted to 6963b8c552c5f7b74a9d2d72e446379a5d8fbec4. I pushed the other patches into https://github.com/certik/symengine/tree/methods5