@certik 2016-01-04
I did push in 6201e785e9519a64e4b52dca6570fe21c48e5b19, since this should be part of this PR. This provides a solid foundation for our future PRs.