@certik 2016-01-04
Btw, here is how to port `StrPrinter` on top of this PR: https://github.com/certik/symengine/commit/1fa371e655d5ef7b9f67f8d723c39792fd9300d5 and all tests pass, except the `MyStrPrinter` ones, because the `bvisit(const Sin &x)` method is never called -- previously I think the `bvisit` binding is determined at compile time of `MyStrPrinter`, while now it is determined at compile time of `StrPrinter`. So the extensibility is limited, but the `StrPrinter` itself works great.