@certik 2016-01-05
AppVeyor got stuck, I created #740 for that.