@isuruf 2016-01-05
See https://github.com/symengine/symengine/commit/98465c6e630cc67f45476df462180981afeed650, which builds on top of your str branch.